Add farming terrain test (#347)
* add terrain test; fix bugs

* change order of move run interactions

* fix clearing metadata

* fix comment

* add exception to runobject in build system for sim terrain

+++ examples/everlon/packages/world/src/systems/MoveSystem.sol
@@ -34,10 +34,15 @@ contract MoveSystem is MoveProtoSystem {
   ) internal override {
     super.postEvent(actingObjectEntityId, objectTypeId, coord, eventEntityId, eventData);
 
-    // Clear all keys in Metadata
-    bytes32[][] memory objectsRan = getKeysInTable(MetadataTableId);
-    for (uint256 i = 0; i < objectsRan.length; i++) {
-      Metadata.deleteRecord(objectsRan[i][0]);
+    address callerAddress = _msgSender();
+    // Clear all keys in Metadata if not called by World or Simulator
+    // This would typically represent the end of a user call, vs the end of
+    // an internal call
+    if (callerAddress != _world() && callerAddress != getSimulatorAddress()) {
+      bytes32[][] memory objectsRan = getKeysInTable(MetadataTableId);
+      for (uint256 i = 0; i < objectsRan.length; i++) {
+        Metadata.deleteRecord(objectsRan[i][0]);
+      }
     }
   }
